Update README.md
This Repository is based on the popular board game called 'Scrabble.' In the game, there are 98 letter tiles with a variation of the 26 letters from the alphabet. Each of the 26 letters has a certain value, and there is a specific number of letters in each tile. This program will take a string of letters (or any case of mixed letters), convert the letters to uppercase, and then returns the scrabble score of the word as an integer whole number. The code must be able to keep track of the supply of letters to know when a value of Zero will be assigned.